+++ Projects/libft/ft_strnew.c
@@ -6,7 +6,7 @@
 /*   By: bchan <marvin@42.fr>                       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/28 13:00:47 by bchan             #+#    #+#             */
-/*   Updated: 2017/11/30 12:59:57 by bchan            ###   ########.fr       */
+/*   Updated: 2017/12/05 11:58:56 by bchan            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,11 +17,11 @@ char	*ft_strnew(size_t size)
 	char	*result;
 	size_t	i;
 
-	result = (char *)malloc(size + 1);
-	i = 0;
+	result = (char *)malloc(size + (size_t)1);
+	i = (size_t)0;
 	if (result)
 	{
-		while (i < size + 1)
+		while (i < size + (size_t)1)
 		{
 			result[i] = '\0';
 			i++;
